backend/data_ingest/requirements.txt
psycopg2-binary>=2.9.10
requests>=2.30.0
python-dateutil>=2.8.2
